requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
PyJWT==2.8.0
google-generativeai==0.4.1
python-jose[cryptography]==3.3.0
jinja2==3.1.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.11
langchain-core==0.1.23
SpeechRecognition==3.10.0
pydub==0.25.1
requests==2.31.0
# python için 3.8 veya üstü gereklidir
